requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
et-xmlfile==1.1.0
futures==3.0.5
idna==3.4
numpy==1.25.1
openpyxl==3.1.2
pandas==2.0.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
stem==1.8.2
tzdata==2023.3
urllib3==2.0.4
